Update and rename HitNote.cs to HitNoteProcessor.cs
